Extract PageHeader to component
