+++ app/utils/utilFunctions.ts
@@ -34,6 +34,7 @@ export const cyrb53 = (str: string, seed = 0) => {
   return 4294967296 * (2097151 & h2) + (h1 >>> 0);
 };
 
+// Ignore / and # for path comparison
 export const normalizePathname = (path: string) => {
   return path.replace(/\/$/, "").split("#")[0];
 };
